Moved ResponseEntityBuilder to util package and added overload methods
